# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Admission Application API",
        "description": "REST API for the Secondary School Admission Registration System. Manages applicant biodata, next-of-kin information, and health records.",
        "contact": {
            "email": "admin@school.edu.ng"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Version 1"
        }
    ],
    "paths": {
...
        "/applicant/dashboard": {
            "get": {
                "tags": [
                    "Applicant Dashboard"
                ],
                "summary": "Get the authenticated applicant's full application",
                "description": "Returns the complete application record — biodata, next-of-kin, health record and current admission status — for the logged-in applicant.",
                "operationId": "2c2dbde65942cea81a24f54e754889c2",
                "responses": {
                    "200": {
                        "description": "Full application record",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApplicantResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — token does not belong to an applicant"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/applicant/status": {
            "get": {
                "tags": [
                    "Applicant Dashboard"
                ],
                "summary": "Get the applicant's admission status",
                "description": "Returns only the application number and current admission status. Useful for a lightweight status-check call.",
                "operationId": "13669f82bd447403852fbca0abc8b564",
                "responses": {
                    "200": {
                        "description": "Admission status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "application_number": {
                                            "type": "string",
                                            "example": "APP-6A36A61045B86"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending",
                                            "enum": [
                                                "pending",
                                                "under_review",
                                                "approved",
                                                "rejected"
                                            ]
                                        },
                                        "submitted_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — token does not belong to an applicant"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/applicant/passport": {
            "post": {
                "tags": [
                    "Applicant Dashboard"
                ],
                "summary": "Upload or replace passport photo",
                "description": "Uploads a passport photo for the authenticated applicant. Send as `multipart/form-data`. Replaces any previously uploaded photo.",
                "operationId": "55ba68dd1022f6854087f77c5d7cdea7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "passport"
                                ],
                                "properties": {
                                    "passport": {
                                        "description": "Image file (jpg, jpeg, png). Max 2 MB.",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Passport uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Passport uploaded successfully."
                                        },
                                        "passport_url": {
                                            "type": "string",
                                            "example": "http://localhost/storage/passports/1/photo.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — token does not belong to an applicant"
                    },
                    "422": {
                        "description": "Validation error (invalid file type or size)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/applicant/print": {
            "get": {
                "tags": [
                    "Applicant Dashboard"
                ],
                "summary": "Get print-ready data for the authenticated applicant's form",
                "description": "Returns a structured payload with all information needed to render and print the admission application form for the logged-in applicant.",
                "operationId": "e84b203528b87161c77e71439b3f9f57",
                "responses": {
                    "200": {
                        "description": "Print-ready application data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "print_data": {
                                            "properties": {
                                                "application_number": {
                                                    "type": "string",
                                                    "example": "APP-6672A1B2C3D4"
                                                },
                                                "submitted_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "passport_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8000/storage/passports/1/photo.jpg",
                                                    "nullable": true
                                                },
                                                "passport_data_uri": {
                                                    "description": "Base64 data URI of the passport photo. Use this for printing/PDF so the image is never blocked by CORS.",
                                                    "type": "string",
                                                    "example": "data:image/jpeg;base64,/9j/4AAQSkZJRg...",
                                                    "nullable": true
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending",
                                                    "enum": [
                                                        "pending",
                                                        "under_review",
                                                        "approved",
                                                        "rejected"
                                                    ]
                                                },
                                                "biodata": {
                                                    "properties": {
                                                        "full_name": {
                                                            "type": "string",
                                                            "example": "Chukwuemeka James Okafor"
                                                        },
                                                        "gender": {
                                                            "type": "string",
                                                            "example": "Male"
                                                        },
                                                        "phone_number": {
                                                            "type": "string",
                                                            "example": "08012345678"
                                                        },
                                                        "home_address": {
                                                            "type": "string",
                                                            "example": "12 Aba Road, Owerri"
                                                        },
                                                        "state": {
                                                            "type": "string",
                                                            "example": "Imo"
                                                        },
                                                        "local_government": {
                                                            "type": "string",
                                                            "example": "Owerri North"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "next_of_kin": {
                                                    "$ref": "#/components/schemas/NextOfKinInput"
                                                },
                                                "health_record": {
                                                    "$ref": "#/components/schemas/HealthRecordInput"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden — token does not belong to an applicant"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NextOfKinInput": {
                "required": [
                    "name",
                    "relationship",
                    "phone_number",
                    "home_address"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Ngozi Okafor"
                    },
                    "relationship": {
                        "type": "string",
                        "example": "Mother",
                        "enum": [
                            "Father",
                            "Mother",
                            "Guardian",
                            "Sibling",
                            "Spouse",
                            "Other"
                        ]
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "08098765432"
                    },
                    "home_address": {
                        "type": "string",
                        "example": "12 Aba Road, Owerri"
                    }
                },
                "type": "object"
            },
            "HealthRecordInput": {
                "required": [
                    "blood_group",
                    "genotype"
                ],
                "properties": {
                    "blood_group": {
                        "type": "string",
                        "example": "O+",
                        "enum": [
                            "A+",
                            "A-",
                            "B+",
                            "B-",
                            "AB+",
                            "AB-",
                            "O+",
                            "O-"
                        ]
                    },
                    "genotype": {
                        "type": "string",
                        "example": "AA",
                        "enum": [
                            "AA",
                            "AS",
                            "AC",
                            "SS",
                            "SC",
                            "CC"
                        ]
                    },
                    "ailments": {
                        "type": "string",
                        "example": "Mild asthma",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApplicantResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "application_number": {
                        "type": "string",
                        "example": "APP-6672A1B2C3D4"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Chukwuemeka"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Okafor"
                    },
                    "middle_name": {
                        "type": "string",
                        "example": "James",
                        "nullable": true
                    },
                    "gender": {
                        "type": "string",
                        "example": "Male",
                        "enum": [
                            "Male",
                            "Female"
                        ]
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "08012345678"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "chukwuemeka@example.com"
                    },
                    "passport_url": {
                        "type": "string",
                        "example": "http://localhost/storage/passports/1/photo.jpg",
                        "nullable": true
                    },
                    "home_address": {
                        "type": "string",
                        "example": "12 Aba Road, Owerri"
                    },
                    "state": {
                        "type": "string",
                        "example": "Imo"
                    },
                    "local_government": {
                        "type": "string",
                        "example": "Owerri North"
                    },
                    "status": {
                        "type": "string",
                        "example": "pending",
                        "enum": [
                            "pending",
                            "under_review",
                            "approved",
                            "rejected"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "next_of_kin": {
                        "$ref": "#/components/schemas/NextOfKinInput"
                    },
                    "health_record": {
                        "$ref": "#/components/schemas/HealthRecordInput"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token returned by POST /api/v1/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}